refactor(markov): make parseAndGetCommand return runner function and slim bin
